Added rollup bundling.
