+++ .editorconfig
@@ -1,0 +1,6 @@
+[*]
+  indent_style = space
+  indent_size = 2
+
+[*.js]
+  quote_type = single
