+++ example/index.html
@@ -7,7 +7,7 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>CJSS</title>
   <link rel="stylesheet" href="style.css">
-  <script src="../cjss.js"></script>
+  <script src="../dist/cjss.min.js"></script>
 </head>
 
 <body>
