+++ src/registerPlugin.js
@@ -1,0 +1,6 @@
+/**
+ * This is just dead code for now
+ **/
+export default function registerPlugin(propertyName, method) {
+
+}
